% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlsmcfcs.R
\name{parlsmcfcs}
\alias{parlsmcfcs}
\title{Run smcfcs in parallel}
\usage{
parlsmcfcs(
  seed = NULL,
  m = 5,
  m_per_core = NULL,
  n_cores = parallel::detectCores() - 1,
  cl_type = "PSOCK",
  outfile = "",
  ...
)
}
\arguments{
\item{seed}{Optional seed}

\item{m}{Number of imputed datasets}

\item{m_per_core}{Optional integer of number of imputations per core.
Default is \code{floor(m / n_cores)}, which it cannot be larger than.}

\item{n_cores}{Number of cores over which to split the m imputations}

\item{cl_type}{Either "PSOCK" or "FORK"}

\item{outfile}{Optional character path to location for
output from the workers. Useful to diagnose rejection sampling warnings.
File path must end with output-filenames.txt}

\item{...}{Additional arguments to pass on to smcfcs::smcfcs}
}
\value{
An object of type "smcfcs"
}
\description{
Run smcfcs in parallel
}
\examples{
# Example here later...
}
